Online supremarket
a program for an online supermarket which includes a website and a server. all product information (name, picture, price, description etc.) is stored in a data base. the supermarket website has multiple tabs, including a checkout, a menu and a basket and a search tab. you can search for pruducts using the category (eg dairy products), name of the product (eg milk) or shortcuts (eg mi for milk or app for apple). for 18+ products you have to go through a "vertification".