a timer with hourglass animation
the layout is created using tkinter and the clock animation is drew using turtle. you can enter time in minutes and press the 'run' button to start the clock. minimal time is 1 minute, when entering a number less then that, the timer will be set for 1 minute automatically. to quit press the quit button. the countdown bwgins as soon as the clock stops rotating. for additional information there is the 'info' button. after the time has ran out a dialogue wingow will appear. you can start the timer again after that.